// types/medrecord/refractometer.ts
export type AutorefractometerData = {
    rightEye: {
      sphere: number;
      cylinder: number;
      axis: number;
    };
    leftEye: {
      sphere: number;
      cylinder: number;
      axis: number;
    };
  };
  
  export type LensometryData = {
    rightEye: {
      sphere: number;
      cylinder: number;
      axis: number;
    };
    leftEye: {
      sphere: number;
      cylinder: number;
      axis: number;
    };
  };